split include/src [WIP]
