+++ include/scaluq/circuit/circuit.cpp
