+++ include/scaluq/constant.hpp
