+++ include/scaluq/gate/gate_pauli.hpp
