+++ include/scaluq/gate/merge_gate.cpp
