+++ include/scaluq/gate/merge_gate.hpp
