+++ include/scaluq/gate/param_gate_pauli.hpp
