+++ include/scaluq/gate/update_ops_dense_matrix.cpp
