+++ include/scaluq/kokkos.hpp
@@ -1,0 +1,8 @@
+#pragma once
+
+namespace scaluq {
+inline void initialize();
+inline void finalize();
+inline bool is_initialized();
+inline bool is_finalized();
+}  // namespace scaluq
