+++ include/scaluq/operator/apply_pauli.hpp
