+++ include/scaluq/operator/operator.cpp
