+++ include/scaluq/operator/operator.hpp
