+++ include/scaluq/operator/pauli_operator.cpp
