+++ include/scaluq/operator/pauli_operator.hpp
