+++ include/scaluq/state/state_vector_batched.cpp
